third_party/blink/web_tests/fast/dom/NodeIterator/NodeIterator-dont-overcollect-expected.txt

PASS nodeFilterObservation.wasCollected is false
PASS callbackObservation.wasCollected is false
PASS callbackWasTriggered is true
PASS nodeIteratorObservation.wasCollected is true
PASS nodeFilterObservation.wasCollected is true
PASS callbackObservation.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
